Replace OrderedDicts with insertion ordered dicts
As of Python 3.6 dictionaries are ordered in insertion order.
Reference: https://docs.python.org/3/library/stdtypes.html?highlight=dict#dict.values

Using older versions of Python will change the order but that is not
problematic as JSON order does not matter in the frontend.

On the downside, the REGIONS dict in views.py can cause differences
when rendering the HTML template (varying region group order).